db/state/execctx: move the rpc-cache tests to a sibling execctx_test pkg
Also name the aura tests package auratests.
