+++ db/state/execctx_test/statecache_rpc_integration_test.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Erigon. If not, see <http://www.gnu.org/licenses/>.
 
-package execctxtests
+package execctx_test
 
 import (
 	"testing"
