+++ execution/protocol/rules/aura/tests/aura_test.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Erigon. If not, see <http://www.gnu.org/licenses/>.
 
-package executiontests
+package auratests
 
 import (
 	"context"
